[FIX] purchase: normalize domain filter
If the Orders domain was combined in a or statement with the quotation
domain, we get, depending on the order, this:
('|',('state','=','draft'),('state','<>','draft'),('state','<>','cancel'))

or this:

('|',('state','<>','draft'),('state','<>','cancel'),('state','=','draft'))

Problem is that those domains do not describe the same set.  Normalizing
the domain fixes the issue.  A complete fix is to correctly normalize
the domain all the time, but it is a much larger work.